[Docker] Disable STRICT_TRANS_TABLES in development database
As of MariaDB >= 10.2.4, the default SQL mode was extended to include
STRICT_TRANS_TABLES and ERROR_FOR_DIVISION_BY_ZERO. This causes problems
when loading member data if ENUM fields are set to an empty string which
is what `script/load-people` defaults to for the `entered_reason` and
`left_reason` columns.

We could switch the default explicitly to 'uknown', which is the default
for the fields in question, but I'm not sure whether this would have any
other knock-on effects - currently any rows inserted with empty strings
for these columns in the existing database have the special empty string
as their value with an index of 0, rather than the index for `unknown`,
which as the default and first value of the ENUM has an index of 1.

We may also want to consider the consequences of this when updating the
production database to a new version - the switch to STRICT_TRANS_TABLES
could catch us out here, too, so we'll need to decide whether to disable
this mode in production going forward or update the existing columns to
conform.

Noting that I've left ERROR_FOR_DIVISION_BY_ZERO enabled for now; I
think it's better to move with the defaults where possible.

See:
 - https://mariadb.com/kb/en/enum/
 - https://mariadb.com/kb/en/sql-mode/